feat: Move user to providers
